# src/app/api/push/subscribe/route.ts
import { NextRequest, NextResponse } from "next/server";
import { upsertSubscription, removeSubscription } from "@/lib/server/subscriptionDb";
import { isSecureSubscriptionEndpoint, normalizeText } from "@/lib/server/pushRouteUtils";
import type { PushSubscriptionRecord } from "@/types";

export const runtime = "nodejs";

function parseSubscription(body: unknown): {
  endpoint: string;
  p256dh: string;
  auth: string;
} | null {
  if (typeof body !== "object" || body === null) return null;

  const subscription = (body as { subscription?: PushSubscriptionJSON }).subscription;
  const endpoint = normalizeText(subscription?.endpoint);
  const p256dh = normalizeText(subscription?.keys?.p256dh);
  const auth = normalizeText(subscription?.keys?.auth);

  if (!endpoint || !p256dh || !auth) return null;
  return { endpoint, p256dh, auth };
}

/**
 * POST /api/push/subscribe
 * Speichert eine neue PushSubscription in der lokalen Subscription-DB.
 *
 * Body: {
 *   subscription: PushSubscriptionJSON,
 *   org: string,
 *   project: string,
 *   azureUserId: string,
 *   displayName: string,
 * }
 */
export async function POST(req: NextRequest): Promise<NextResponse> {
  let body: unknown;

  try {
    body = await req.json();
  } catch {
    return NextResponse.json({ error: "Ungueltiger Request-Body" }, { status: 400 });
  }

  const subscription = parseSubscription(body);
  const org = normalizeText((body as { org?: string })?.org);
  const project = normalizeText((body as { project?: string })?.project);
  const azureUserId = normalizeText((body as { azureUserId?: string })?.azureUserId);
  const displayName = normalizeText((body as { displayName?: string })?.displayName);

  if (!subscription) {
    return NextResponse.json({ error: "Unvollstaendige Subscription-Daten" }, { status: 400 });
  }
  if (!isSecureSubscriptionEndpoint(subscription.endpoint)) {
    return NextResponse.json({ error: "Ungueltiger Subscription-Endpoint" }, { status: 400 });
  }

  if (!org || !project || !azureUserId) {
    return NextResponse.json(
      { error: "org, project und azureUserId sind Pflichtfelder" },
      { status: 400 }
    );
  }

  const record: PushSubscriptionRecord = {
    endpoint: subscription.endpoint,
    keys: {
      p256dh: subscription.p256dh,
      auth: subscription.auth,
    },
    org,
    project,
    azureUserId,
    displayName: displayName || "Unbekannter Benutzer",
    createdAt: new Date().toISOString(),
  };

  try {
    upsertSubscription(record);
  } catch (err) {
    console.error("[push/subscribe] Fehler beim Speichern der Subscription:", err);
    return NextResponse.json({ error: "Speicherfehler" }, { status: 500 });
  }

  return NextResponse.json({ ok: true }, { status: 201 });
}

/**
 * DELETE /api/push/subscribe
 * Loescht eine PushSubscription anhand ihres Endpoints.
 *
 * Body: { endpoint: string }
 */
export async function DELETE(req: NextRequest): Promise<NextResponse> {
  let body: unknown;

  try {
    body = await req.json();
  } catch {
    return NextResponse.json({ error: "Ungueltiger Request-Body" }, { status: 400 });
  }

  const endpoint = normalizeText((body as { endpoint?: string })?.endpoint);

  if (!endpoint) {
    return NextResponse.json({ error: "endpoint ist ein Pflichtfeld" }, { status: 400 });
  }

  try {
    removeSubscription(endpoint);
  } catch (err) {
    console.error("[push/subscribe] Fehler beim Loeschen der Subscription:", err);
    return NextResponse.json({ error: "Loeschfehler" }, { status: 500 });
  }

  return NextResponse.json({ ok: true });
}